Made a VandyResult abstraction for multi-sport analysis, applied it to football, refactored football to its own place
